+++ src/Frontend/Core/Engine/Exception.php
@@ -5,6 +5,6 @@ namespace Frontend\Core\Engine;
 /**
  * This class is used to handle frontend related exceptions.
  */
-class Exception extends \SpoonException
+class Exception extends \RuntimeException
 {
 }